requirements.txt
# Core dependencies for Movie Recommendation System
streamlit>=1.28.0
tmdbv3api>=1.9.0
requests>=2.31.0

# Natural Language Processing
nltk>=3.8.0
textstat>=0.7.3

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning & Embeddings
sentence-transformers>=2.2.0
torch>=2.0.0
scikit-learn>=1.3.0

# Google Sheets Integration
gspread>=5.10.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Optional: For better testing
coverage>=7.2.0
flake8>=6.0.0
black>=23.0.0

# Utility
python-dotenv>=1.0.0